Backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.0.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
facenet-pytorch==2.5.3
filelock==3.12.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.3
mpmath==1.3.0
mysqlclient==2.1.1
networkx==3.1
numpy==1.24.3
Pillow==9.5.0
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.1.0
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
